requirements.txt
# Deep Learning Framework (compatible with Python 3.12)
tensorflow>=2.16.0

# Data Handling & Loading
datasets>=2.14.0
huggingface-hub>=0.16.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Progress Tracking
tqdm>=4.65.0

# Web Application
streamlit>=1.25.0

# Jupyter Notebook Support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Development Tools (optional)
black>=23.7.0
flake8>=6.1.0
pytest>=7.4.0
